Fix record formatting
