+++ baremaps-core/src/main/java/org/apache/baremaps/database/metadata/ColumnResult.java
@@ -13,26 +13,27 @@
 package org.apache.baremaps.database.metadata;
 
 public record ColumnResult(
-  String tableCat,
-  String tableSchem,
-  String tableName,
-  String columnName,
-  int dataType,
-  String typeName,
-  int columnSize,
-  int decimalDigits,
-  int numPrecRadix,
-  int nullable,
-  String remarks,
-  String columnDef,
-  int sqlDataType,
-  int sqlDateTimeSub,
-  int charOctetLenght,
-  int ordinalPosition,
-  String isNullable,
-  String scopeCatalog,
-  String scopeSchema,
-  String scopeTable,
-  short sourceDataType,
-  String isAutoIncrement,
-  String isGeneratedColumn) {}
+    String tableCat,
+    String tableSchem,
+    String tableName,
+    String columnName,
+    int dataType,
+    String typeName,
+    int columnSize,
+    int decimalDigits,
+    int numPrecRadix,
+    int nullable,
+    String remarks,
+    String columnDef,
+    int sqlDataType,
+    int sqlDateTimeSub,
+    int charOctetLenght,
+    int ordinalPosition,
+    String isNullable,
+    String scopeCatalog,
+    String scopeSchema,
+    String scopeTable,
+    short sourceDataType,
+    String isAutoIncrement,
+    String isGeneratedColumn) {
+}
