+++ baremaps-core/src/main/java/org/apache/baremaps/database/metadata/TableResult.java
@@ -13,13 +13,14 @@
 package org.apache.baremaps.database.metadata;
 
 public record TableResult(
-  String tableCat,
-  String tableSchem,
-  String tableName,
-  String tableType,
-  String remarks,
-  String typeCat,
-  String typeSchem,
-  String typeName,
-  String selfReferencingColName,
-  String refGeneration) {}
+    String tableCat,
+    String tableSchem,
+    String tableName,
+    String tableType,
+    String remarks,
+    String typeCat,
+    String typeSchem,
+    String typeName,
+    String selfReferencingColName,
+    String refGeneration) {
+}
