+++ baremaps-core/src/main/java/org/apache/baremaps/stream/ConsumerUtils.java
@@ -24,7 +24,7 @@ public class ConsumerUtils {
 
   /**
    * Returns a consumer that applies a function to its input, and then passes the result to the
-   * 
+   *
    * @param type the type of the input
    * @return
    * @param <T>
