+++ baremaps-core/src/main/java/org/apache/baremaps/workflow/tasks/ExecuteSql.java
@@ -12,16 +12,13 @@
 
 package org.apache.baremaps.workflow.tasks;
 
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.sql.SQLException;
+import java.util.Arrays;
 import org.apache.baremaps.workflow.Task;
 import org.apache.baremaps.workflow.WorkflowContext;
 import org.apache.baremaps.workflow.WorkflowException;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.sql.SQLException;
-import java.util.Arrays;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,16 +34,16 @@ public record ExecuteSql(String database, Path file, boolean parallel) implement
       queries = queries.parallel();
     }
     queries.forEach(
-      query -> {
-        var dataSource = context.getDataSource(database);
-        try (var connection = dataSource.getConnection()) {
-          connection.createStatement().execute(query);
-        } catch (SQLException e) {
-          logger.error("Failed executing {}", query);
-          throw new WorkflowException(e);
-        }
-      });
+        query -> {
+          var dataSource = context.getDataSource(database);
+          try (var connection = dataSource.getConnection()) {
+            connection.createStatement().execute(query);
+          } catch (SQLException e) {
+            logger.error("Failed executing {}", query);
+            throw new WorkflowException(e);
+          }
+        });
     logger.info("Finished executing {}", file);
   }
-  
+
 }
